requirements.txt
fastapi
uvicorn[standard]
supabase
pydantic
python-dotenv
openai
crewai
langchain-openai